Added type to checkout
