+++ src/components/directory/directory.styles.tsx
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 
 export const DirectoryContainer = styled.div`
-  padding: 0 40px;
-  margin: 0;
+	padding: 0 40px;
+	margin: 0;
 
-  @media (max-width: 480px) {
-    padding: 0;
-  }
-`
+	@media (max-width: 480px) {
+		padding: 0;
+	}
+`;
